Adds endpoints for deals

+++ spec/support/shared_contexts/with_deal.rb
@@ -1,18 +1,20 @@
 # frozen_string_literal: true
 
 RSpec.shared_context 'with deal params', :with_deal_params do
-  include_context "with existing pipeline"
+  include_context 'with existing pipeline'
+  include_context 'with existing contact'
+
   let(:deal_name)        { 'Bogus Deal' }
   let(:deal_deal_url)    { 'https://www.mhenrixon.com/bogus' }
   let(:deal_title)       { 'Bogus Deal' }
   let(:deal_description) { 'This deal is completely bogus' }
   let(:deal_deal)        { 'bogus-deal' }
-  let(:deal_contact)     {}
+  let(:deal_contact)     { contact_id }
   let(:deal_value)       { 156.24 }
   let(:deal_currency)    { 'eur' }
-  let(:deal_group)       {}
-  let(:deal_stage)       { pipeline_id }
-  let(:deal_owner)       {}
+  let(:deal_group)       { pipeline_id }
+  let(:deal_stage)       {}
+  let(:deal_owner)       { user_id }
   let(:deal_percent)     { 50 }
   let(:deal_status)      { 0 }
   let(:deal_params) do
@@ -45,9 +47,6 @@ RSpec.shared_context 'with existing deal', with_existing_deal: true do
   let(:deal_id) { deal[:id] }
 
   after do
-    if deal_id
-      client.update_deal(deal_id)
-      client.delete_deal(deal_id)
-    end
+    client.delete_deal(deal_id) if deal_id
   end
 end
